created queries for chat
